requirements.txt
numpy>=1.24.3
pandas>=2.0.3
ccxt>=4.0.0
gymnasium>=0.26.2
stable-baselines3>=2.1.0
ray[rllib]>=2.7.0
streamlit>=1.26.0
fastapi>=0.103.1
uvicorn>=0.23.2
mlflow>=2.7.1
python-dotenv>=1.0.0
pytest>=7.4.2
pyyaml>=6.0.1
pyqlib>=0.9.5
plotly>=5.18.0
ta-lib>=0.4.26
scikit-learn>=1.3.0
requests>=2.31.0